backend/requirements.txt
yfinance>=0.2.40
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0
matplotlib>=3.8.0
plotly>=5.20.0
streamlit>=1.32.0
openai>=1.25.0
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.8.0